requirements.txt
Django==1.11.16
# Libs we want installed first so that later we regard them as installed
-e git+https://github.com/open-contracting/lib-cove-ocds.git@f705ae7ec729eb094b8c06a0dc9ec8dfae438554#egg=libcoveocds
# Libs we need
-e git+https://github.com/OpenDataServices/flatten-tool.git@4c13ef0b32a59e810919a3de09bc8f64ce8f9392#egg=flattentool
-e git+https://github.com/OpenDataServices/lib-cove.git@af8caaf0643fee72c697a41592865183ad55c4b5#egg=libcove
-e git+https://github.com/OpenDataServices/cove.git@9d2145debe25bb55e67887de2ae7e22e4cf5467d#egg=cove
-e git+https://github.com/odscjames/openownership-lib-cove-bods-alpha.git@24b258d2773e12f990bd8d2b88e68b1183d1f96d#egg=libcovebods
## The following requirements were added by pip freeze:
bleach==3.0.2
cached-property==1.5.1
certifi==2018.1.18
chardet==3.0.4
commonmark==0.8.1
dealer==2.0.5
django-bootstrap3==9.1.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
et-xmlfile==1.0.1
future==0.16.0
html5lib==1.0.1
idna==2.6
jdcal==1.3
json-merge-patch==0.2
jsonref==0.2
jsonschema==2.6.0
LEPL==5.1.3
lxml==4.2.5
openpyxl==2.5.10
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.7
raven==6.6.0
requests==2.20.1
rfc3987==1.3.7
rfc6266==0.0.4
schema==0.6.8
six==1.11.0
sqlparse==0.2.4
strict-rfc3339==0.7
urllib3==1.22
webencodings==0.5.1
xmltodict==0.11.0
